+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+## [0.4.2] - 2023-03-31
+
+### Added
+
+- Adds a NativeResponseHandler to abstractions. 
+- Adds setResponseHandler method to RequestInformation class in abstractions. 
+
 ## [0.4.1] - 2023-03-29
 
 ### Changed